Add multiple ingredient/instruction prompt. must refactor

+++ yumby/Ingredient.cs
@@ -3,7 +3,7 @@
 public class Ingredient
 {
     public string? Name { get; set; }
-    public double? Quantity { get; set; }
+    public decimal? Quantity { get; set; }
     public string? Unit { get; set; }
     //nutritional info?
 }
